coloquei o valor monet[ario de caixa em bigdecimal e arrumei um bug ainda bem que o vi

+++ barbearia-backend/dominio-principal/src/test/java/com/cesarschool/cucumber/gestaoCaixa/infraestrutura/LancamentoFactory.java
@@ -1,5 +1,7 @@
 package com.cesarschool.cucumber.gestaoCaixa.infraestrutura;
 
+import java.math.BigDecimal;
+
 import com.cesarschool.barbearia.dominio.principal.cliente.ClienteId;
 import com.cesarschool.barbearia.dominio.principal.cliente.caixa.Lancamento;
 import com.cesarschool.barbearia.dominio.principal.cliente.caixa.MeioPagamento;
@@ -8,12 +10,12 @@ public final class LancamentoFactory {
     private LancamentoFactory() {}
 
     public static Lancamento entrada(String desc, double valor) { 
-        return Lancamento.novoRecibemento(desc, valor, MeioPagamento.DINHEIRO); 
+        return Lancamento.novoRecibemento(desc, BigDecimal.valueOf(valor), MeioPagamento.DINHEIRO); 
     }
     public static Lancamento saida(String desc, double valor) { 
-        return Lancamento.novoGasto(desc, valor, MeioPagamento.DINHEIRO); 
+        return Lancamento.novoGasto(desc, BigDecimal.valueOf(valor), MeioPagamento.DINHEIRO); 
     }
     public static Lancamento pendente(ClienteId cliente, String desc, double valor) { 
-        return Lancamento.novaDivida(cliente, desc, valor, MeioPagamento.DINHEIRO); 
+        return Lancamento.novaDivida(cliente, desc, BigDecimal.valueOf(valor), MeioPagamento.DINHEIRO); 
     }
 }
